requirements.txt
Flask
google-cloud-datastore
google-api-python-client>=1.7.10
google-auth>=1.11.2
google-auth-httplib2
google-auth-oauthlib>=0.4.1
httplib2>=0.17.0
pytz
